# go-kratos/api/gen/pbauth/register.pb.go
// Code generated by protoc-gen-go. DO NOT EDIT.
// versions:
// 	protoc-gen-go v1.36.6
// 	protoc        v6.31.1
// source: proto/pbauth/register.proto

package pbauth

import (
	pbuser "go-kratos/api/gen/pbuser"
	protoreflect "google.golang.org/protobuf/reflect/protoreflect"
	protoimpl "google.golang.org/protobuf/runtime/protoimpl"
	reflect "reflect"
	sync "sync"
	unsafe "unsafe"
)

const (
	// Verify that this generated code is sufficiently up-to-date.
	_ = protoimpl.EnforceVersion(20 - protoimpl.MinVersion)
	// Verify that runtime/protoimpl is sufficiently up-to-date.
	_ = protoimpl.EnforceVersion(protoimpl.MaxVersion - 20)
)

// 请求-注册
type RegisterRequest struct {
	state protoimpl.MessageState `protogen:"open.v1"`
	// 昵称
	Nickname string `protobuf:"bytes,1,opt,name=nickname,proto3" json:"nickname,omitempty"`
	// 邮箱
	Email string `protobuf:"bytes,2,opt,name=email,proto3" json:"email,omitempty"`
	// 密码
	Password      string `protobuf:"bytes,3,opt,name=password,proto3" json:"password,omitempty"`
	unknownFields protoimpl.UnknownFields
	sizeCache     protoimpl.SizeCache
}

func (x *RegisterRequest) Reset() {
	*x = RegisterRequest{}
	mi := &file_proto_pbauth_register_proto_msgTypes[0]
	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
	ms.StoreMessageInfo(mi)
}

func (x *RegisterRequest) String() string {
	return protoimpl.X.MessageStringOf(x)
}

func (*RegisterRequest) ProtoMessage() {}

func (x *RegisterRequest) ProtoReflect() protoreflect.Message {
	mi := &file_proto_pbauth_register_proto_msgTypes[0]
	if x != nil {
		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
		if ms.LoadMessageInfo() == nil {
			ms.StoreMessageInfo(mi)
		}
		return ms
	}
	return mi.MessageOf(x)
}

// Deprecated: Use RegisterRequest.ProtoReflect.Descriptor instead.
func (*RegisterRequest) Descriptor() ([]byte, []int) {
	return file_proto_pbauth_register_proto_rawDescGZIP(), []int{0}
}

func (x *RegisterRequest) GetNickname() string {
	if x != nil {
		return x.Nickname
	}
	return ""
}

func (x *RegisterRequest) GetEmail() string {
	if x != nil {
		return x.Email
	}
	return ""
}

func (x *RegisterRequest) GetPassword() string {
	if x != nil {
		return x.Password
	}
	return ""
}

// 响应-注册
type RegisterResponse struct {
	state         protoimpl.MessageState `protogen:"open.v1"`
	Token         string                 `protobuf:"bytes,1,opt,name=token,proto3" json:"token,omitempty"`
	User          *pbuser.User           `protobuf:"bytes,2,opt,name=user,proto3" json:"user,omitempty"`
	unknownFields protoimpl.UnknownFields
	sizeCache     protoimpl.SizeCache
}

func (x *RegisterResponse) Reset() {
	*x = RegisterResponse{}
	mi := &file_proto_pbauth_register_proto_msgTypes[1]
	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
	ms.StoreMessageInfo(mi)
}

func (x *RegisterResponse) String() string {
	return protoimpl.X.MessageStringOf(x)
}

func (*RegisterResponse) ProtoMessage() {}

func (x *RegisterResponse) ProtoReflect() protoreflect.Message {
	mi := &file_proto_pbauth_register_proto_msgTypes[1]
	if x != nil {
		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
		if ms.LoadMessageInfo() == nil {
			ms.StoreMessageInfo(mi)
		}
		return ms
	}
	return mi.MessageOf(x)
}

// Deprecated: Use RegisterResponse.ProtoReflect.Descriptor instead.
func (*RegisterResponse) Descriptor() ([]byte, []int) {
	return file_proto_pbauth_register_proto_rawDescGZIP(), []int{1}
}

func (x *RegisterResponse) GetToken() string {
	if x != nil {
		return x.Token
	}
	return ""
}

func (x *RegisterResponse) GetUser() *pbuser.User {
	if x != nil {
		return x.User
	}
	return nil
}

var File_proto_pbauth_register_proto protoreflect.FileDescriptor

const file_proto_pbauth_register_proto_rawDesc = "" +
	"\n" +
	"\x1bproto/pbauth/register.proto\x12\x06pbauth\x1a\x17proto/pbuser/user.proto\"_\n" +
	"\x0fRegisterRequest\x12\x1a\n" +
	"\bnickname\x18\x01 \x01(\tR\bnickname\x12\x14\n" +
	"\x05email\x18\x02 \x01(\tR\x05email\x12\x1a\n" +
	"\bpassword\x18\x03 \x01(\tR\bpassword\"J\n" +
	"\x10RegisterResponse\x12\x14\n" +
	"\x05token\x18\x01 \x01(\tR\x05token\x12 \n" +
	"\x04user\x18\x02 \x01(\v2\f.pbuser.UserR\x04userB!Z\x1fgo-kratos/api/gen/pbauth;pbauthb\x06proto3"

var (
	file_proto_pbauth_register_proto_rawDescOnce sync.Once
	file_proto_pbauth_register_proto_rawDescData []byte
)

func file_proto_pbauth_register_proto_rawDescGZIP() []byte {
	file_proto_pbauth_register_proto_rawDescOnce.Do(func() {
		file_proto_pbauth_register_proto_rawDescData = protoimpl.X.CompressGZIP(unsafe.Slice(unsafe.StringData(file_proto_pbauth_register_proto_rawDesc), len(file_proto_pbauth_register_proto_rawDesc)))
	})
	return file_proto_pbauth_register_proto_rawDescData
}

var file_proto_pbauth_register_proto_msgTypes = make([]protoimpl.MessageInfo, 2)
var file_proto_pbauth_register_proto_goTypes = []any{
	(*RegisterRequest)(nil),  // 0: pbauth.RegisterRequest
	(*RegisterResponse)(nil), // 1: pbauth.RegisterResponse
	(*pbuser.User)(nil),      // 2: pbuser.User
}
var file_proto_pbauth_register_proto_depIdxs = []int32{
	2, // 0: pbauth.RegisterResponse.user:type_name -> pbuser.User
	1, // [1:1] is the sub-list for method output_type
	1, // [1:1] is the sub-list for method input_type
	1, // [1:1] is the sub-list for extension type_name
	1, // [1:1] is the sub-list for extension extendee
	0, // [0:1] is the sub-list for field type_name
}

func init() { file_proto_pbauth_register_proto_init() }
func file_proto_pbauth_register_proto_init() {
	if File_proto_pbauth_register_proto != nil {
		return
	}
	type x struct{}
	out := protoimpl.TypeBuilder{
		File: protoimpl.DescBuilder{
			GoPackagePath: reflect.TypeOf(x{}).PkgPath(),
			RawDescriptor: unsafe.Slice(unsafe.StringData(file_proto_pbauth_register_proto_rawDesc), len(file_proto_pbauth_register_proto_rawDesc)),
			NumEnums:      0,
			NumMessages:   2,
			NumExtensions: 0,
			NumServices:   0,
		},
		GoTypes:           file_proto_pbauth_register_proto_goTypes,
		DependencyIndexes: file_proto_pbauth_register_proto_depIdxs,
		MessageInfos:      file_proto_pbauth_register_proto_msgTypes,
	}.Build()
	File_proto_pbauth_register_proto = out.File
	file_proto_pbauth_register_proto_goTypes = nil
	file_proto_pbauth_register_proto_depIdxs = nil
}